+++ bigquery/CHANGES.md
@@ -13,6 +13,8 @@
         mapred.bq.bigquery.root.url
         mapred.bq.token.server.url
 
+1.  Use Slf4j backend by default for Google Flogger.
+
 ### 1.0.1 - 2020-02-13
 
 1.  Fix shaded jar - add back missing relocated dependencies.