+++ gcs/CHANGES.md
@@ -30,6 +30,8 @@
 
 1.  Always use URI path encoding and remove `fs.gs.path.encoding` property.
 
+1.  Use Slf4j backend by default for Google Flogger.
+
 ### 2.0.1 - 2020-02-13
 
 1.  Cooperative Locking FSCK tool: fix recovery of operations that failed before